# warren/openapi.yml
openapi: 3.0.3
info:
  title: warren API
  version: 0.1.0
  description: |
    Coordination hub for LLM agents. Admin (PSK-authenticated) injects
    requests and memos; agents (Bearer-token-authenticated) claim
    requests, respond, and acknowledge memos.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: warren_session
    agentToken:
      type: http
      scheme: bearer
  schemas:
    Agent:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        class: { type: string }
        type: { type: string, nullable: true }
        model: { type: string }
        authtoken: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, name, class, type, model, authtoken, created_at]
    AgentNew:
      type: object
      properties:
        name: { type: string }
        class: { type: string }
        type: { type: string, nullable: true }
        model: { type: string }
      required: [name, class, model]
    AgentPatch:
      type: object
      properties:
        name: { type: string }
        class: { type: string }
        model: { type: string }
    LoginReq:
      type: object
      properties:
        password: { type: string }
      required: [password]
    LoginRes:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
paths:
  /api/login:
    post:
      summary: Log in as admin (sets warren_session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginReq' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginRes' }
        '401':
          description: Bad PSK
...
  /api/agents:
    get:
      summary: List agents (admin)
      security:
        - adminSession: []
        - agentToken: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Agent' }
    post:
      summary: Create agent (admin); returns server-generated authtoken
      security:
        - adminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentNew' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
...
  /api/agents/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get agent (admin)
      security: [{ adminSession: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '404': { description: Not found }
    put:
      summary: Patch agent (admin)
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
    delete:
      summary: Delete agent (admin)
      security: [{ adminSession: [] }]
      responses:
        '204': { description: OK }
...